Predicting house prices using regression
In the project, a linear regression model is created to predict the house prices starting form the basic features like square feet living and then multiple features are used to predict the house prices. The quality of model is determined by the root mean square (RMSE) error at the end.